pull out multislitflexure class
